Support adding JSON files to index
I've added support to add specific JSON files to an ES index in
upload_tests.py.
I've also patched results.json to be parseable. The problem is
that it doesn't seem to be a proper JSON file. I know it is produced
by a tool, so the problem should be fixed by fixing the tool.
But in any case, you can check the changes needed by checking at
the diff. It is mainly that last elements in a JSON object are not
followed by comma, and that we need value for all properties. I've
used "" for missing values, but in some cases probably it would
be more appropriate to use NaN, which I guess is valid JSON too.